requirements.txt
prometheus-client==0.18.0
requests==2.31.1
python-dotenv = 1.0.0